Move RangeDataEqualsimpl to internal header
